♻️ refactor: 관리자 페이지 경로 수정 및 폴더 구조 정리
